q2_1
Tested and works in python 3.8, sufficiently answers the prompted question in mini-writeup.txt (probably want to move findings from this into a collective document such as the jupyter notebook we discussed). 

The code is not the most efficient but should be easy enough to follow. The graph dimensions are hard coded into python lists, but can be changed easily if the number of epochs we want to iterate for increase/decrease. 
 This script requires both matplotlib.pyplot and numpy to be installed.